requirements.txt
requests
python-dotenv
googletrans==4.0.0-rc1
CurrencyConverter
pandas